Introduce openstack-login and openstack-logout commands
openstack-login: Prompts for a I/D user and password that is used to
read the technical "korifi-dev-bot" user from the openstack secrets.
This user is then set as an v3applicationcredential and the env is
purged of any traces of the SAP user used to obtain the secret. The
technical user details are stored in the zsh profile so that one is not
asked to continuoussly provide their credentials

openstack-logout: Deletes the technical user profile and unsets all
openstack related env vars